# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ResumeAI.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ResumeMetadata": {
      "get": {
        "tags": [
          "ResumeMetadata"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "ResumeMetadata"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeMetadata"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeMetadata"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeMetadata"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ResumeMetadata"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ResumeMetadata/file-info": {
      "patch": {
        "tags": [
          "ResumeMetadata"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResumeFileInfoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResumeFileInfoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResumeFileInfoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResumeMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "blobPath": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "fileSize": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "fileUrl": {
            "type": "string",
            "nullable": true
          },
          "uploadDate": {
            "type": "string",
            "format": "date-time"
          },
          "jobTitle": {
            "type": "string",
            "nullable": true
          },
          "currentCompany": {
            "type": "string",
            "nullable": true
          },
          "yearsOfExperience": {
            "type": "string",
            "nullable": true
          },
          "professionalBio": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateResumeFileInfoRequest": {
        "type": "object",
        "properties": {
          "blobPath": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "fileSize": {
            "type": "integer",
            "format": "int32"
          },
          "fileUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter 'Bearer {token}' in the field below",
        "scheme": "Bearer"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}